requirements.txt
urlextract
urllib3
bs4
cssbeautifier
jsbeautifier
chardet
pywebview
logging
html5lib
pygame
pygame-textinput
pillow
python-magic-bin
psutil
asyncio
pyinstaller